+++ src/dragularService.js
@@ -64,8 +64,6 @@ var dragularService = function ( $rootScope, $compile ) {
 	// service definition
 	function service( arg0, arg1 ) {
 
-		// console.log('dragularService arg0, arg1', arg0, arg1);
-
 		var initialContainers = arg0 || [],
 			options = arg1 || {},
 			o, // shorthand for options
@@ -159,19 +157,17 @@ var dragularService = function ( $rootScope, $compile ) {
 
 		function processServiceArguments() {
 
-			if ( arguments.length === 1 && // if there is only one argument we need to distinguish if it is options object or container(s) reference
-
+			if ( typeof arg0 === 'string' ) {
+				initialContainers = document.querySelectorAll( arg0 );
+			}
+			else if (
+        arguments.length === 1 && // if there is only one argument we need to distinguish if it is options object or container(s) reference
 				!_isArray( arg0 ) && // array of containers elements
 				!angular.isElement( arg0 ) && // one container element
-				!arg0[ 0 ] && // array-like object with containers elements
-				typeof arg0 !== 'string' ) { // selector
+				!arg0[ 0 ]){ // array-like object with containers elements
 				// then arg0 is options object
 				options = arg0 || {};
 				initialContainers = []; // containers are not provided on init
-			}
-			else if ( typeof arg0 === 'string' ) {
-
-				initialContainers = document.querySelectorAll( arg0 );
 			}
 
 			o = options.copyOptions ? angular.copy( options ) : options;
@@ -257,7 +253,7 @@ var dragularService = function ( $rootScope, $compile ) {
 					shared.containersCtx[ nameSpace ].push({
 						o: o,
 						m: getContainersModel( o )[ i ], // can be undefined
-						fm: o.containersFilteredModel[ i ] // can be undefined
+						fm: getContainersFilteredModel( o )[ i ] // can be undefined
 					});
 				}
 			} );
@@ -458,7 +454,7 @@ var dragularService = function ( $rootScope, $compile ) {
 			var containerIndex = getContainers( o ).indexOf( context.source );
 			shared.sourceModel = getContainersModel( o )[ containerIndex ];
 
-			shared.sourceFilteredModel = o.containersFilteredModel[ containerIndex ];
+			shared.sourceFilteredModel = getContainersFilteredModel( o )[ containerIndex ];
 			shared.initialIndex = domIndexOf( context.item, context.source );
 
 			drake.dragging = true;
@@ -637,7 +633,7 @@ var dragularService = function ( $rootScope, $compile ) {
 
 				if ( g( o.removeOnSpill ) === true ) {
 					type === 'over' ? spillOver() : spillOut();
-				}	
+				}
 
 				function notify( scope ){
 					scope.$emit( o.eventNames[ 'dragular' + type ], shared.item, shared.lastDropTarget, shared.source, e );
@@ -801,6 +797,10 @@ var dragularService = function ( $rootScope, $compile ) {
 				return false;
 			}
 
+      if ( getContainers( o ).indexOf( el ) > -1) {
+        return true;
+      }
+
 			var i = o.nameSpace.length;
 			while ( i-- ) {
 
@@ -828,6 +828,11 @@ var dragularService = function ( $rootScope, $compile ) {
 		function getContainersModel( opt ) {
 
 			return _getContainers( 'containersModel', opt, true );
+		}
+
+		function getContainersFilteredModel( opt ) {
+
+			return _getContainers( 'containersFilteredModel', opt, true );
 		}
 
 		function _getContainers( containersType, opt, to2d ) {
@@ -1070,6 +1075,35 @@ var dragularService = function ( $rootScope, $compile ) {
 			release( {} );
 		}
 
+		function addContainers( containers, containersModel, containersFilteredModel ) {
+
+      if (!containers) {
+        return;
+      }
+
+      containersModel = containersModel || getContainersModel( o ) || {};
+      containersFilteredModel = containersFilteredModel || getContainersFilteredModel( o ) || {};
+      containers = _isArray( containers ) ? containers : makeArray( containers );
+      containers.forEach( function forEachContainer( container ) {
+
+        angular.forEach( o.nameSpace, function forEachNs( nameSpace ) {
+          if (shared.containers[ nameSpace ].indexOf( container ) > -1) {
+            return;
+          }
+
+          var newIndex = shared.containers[ nameSpace ].push( container ) - 1;
+          shared.containersCtx[ nameSpace ].push({
+            o: o,
+            m: containersModel[ newIndex ], // can be undefined
+            fm: containersFilteredModel[ newIndex ] // can be undefined
+          });
+
+          regEvent( container, 'off', 'mousedown', grab );
+					regEvent( container, 'on', 'mousedown', grab );
+        } );
+      } );
+    }
+
 		function removeContainers( containers ) {
 
 			$rootScope.$applyAsync( function applyDestroyed() {
@@ -1083,29 +1117,7 @@ var dragularService = function ( $rootScope, $compile ) {
 						index = shared.containers[ nameSpace ].indexOf( container );
 						shared.containers[ nameSpace ].splice( index, 1 );
 						shared.containersCtx[ nameSpace ].splice( index, 1 );
-
-						if ( -1 < (index = initialContainers.indexOf( container ))) {
-							initialContainers.splice( index, 1 );
-						}
 					} );
-				} );
-			} );
-		}
-
-		function addContainers( containers ) {
-
-			containers = _isArray( containers ) ? containers : makeArray( containers );
-			containers.forEach( function forEachContainer( container, i ) {
-
-				angular.forEach( o.nameSpace, function forEachNs( nameSpace ) {
-
-					shared.containers[ nameSpace ].push( container );
-					shared.containersCtx[ nameSpace ].push({
-						o: o,
-						m: getContainersModel( o )[ i ], // can be undefined
-						fm: o.containersFilteredModel[ i ] // can be undefined
-					});
-					initialContainers.push( container );
 				} );
 			} );
 		}